backend/requirements.txt
Flask
flask-cors
python-dotenv
requests
langchain
langchain-google-genai
gunicorn
